==> 02_TwitterDataCleaning.ipynb <==
# Twitter Data Cleaning

from time import sleep
import pandas as pd
import re
from tqdm import tnrange, tqdm_notebook, tqdm

PRICE_FOLDER    = "data/price/" 
TWITTER_FOLDER  = "data/twitter/"
tweets_raw_file = './data/twitter/bitcoin_tweets_raw.csv'
tweets_clean_file = './data/twitter/bitcoin_tweets_clean.csv'

# Read in data
d = pd.read_csv(tweets_raw_file)
# Drop duplicate rows to manage dulpicate headers
d = d[d.ID !='ID']
d = d.reset_index(drop=True)

for i, s in enumerate(tqdm(d['Tweets'])):
    text = d.loc[i, 'Tweets']
    text = text.replace("#", "")
    text = text.replace("@", "")
    text = re.sub('https?://(?:[-\w.]|(?:%[\da-fA-F]{2}))+', '', text, flags=re.MULTILINE)
    text = re.sub('@\\w+ *', '', text, flags=re.MULTILINE)
    d.loc[i, 'Tweets'] = text

# Remove keywords from dataframe
to_remove = ["Airdrop", "airdrop", "freebitcoin", "freebtc"]
d = d[~d.Tweets.str.contains("|".join(to_remove))]
d = d.reset_index(drop=True)

# Manage dataframe times
d['CreatedAt'] = pd.to_datetime(d['CreatedAt'])
twitter_hr = []
twitter_next_hr = []
for i, s in enumerate(tqdm(d['CreatedAt'])):
    t = s.floor(freq='H')
    n = s.ceil(freq='H')
    twitter_hr.append(t)
    twitter_next_hr.append(n)
d['Hour'] = twitter_hr
d['NextHour'] = twitter_next_hr

# Read clean file
try:
    clean_df = pd.read_csv(tweets_clean_file)
except pd.io.common.EmptyDataError:
    clean_df = pd.DataFrame()
# Merge raw and clean
clean_df_updated = pd.concat([clean_df, d])
# Drop duplicates
clean_df_updated = clean_df_updated.drop_duplicates(subset=['Tweets'])

# Write clean csv
f = open(tweets_clean_file, 'a+', encoding='utf-8')
clean_df_updated.to_csv(f, header=True, encoding='utf-8',index=False)
f.close()

# Update and clear raw csv
r = open(tweets_raw_file, 'w+', encoding='utf-8')
r.close()

print('Twitter data cleaned sucessfully')

## Write script

!jupyter nbconvert --to script --no-prompt 02_TwitterDataCleaning.ipynb
