==> 03_TwitterSentimentAnalysis.ipynb <==
# Split files and compute sentiment scores

from time import sleep
import json
import pandas as pd
import io
from tqdm import tqdm

PRICE_FOLDER    = "data/price/"
TWITTER_FOLDER  = "data/twitter/"
tweets_raw_file = './data/twitter/bitcoin_tweets_raw.csv'
tweets_clean_file = './data/twitter/bitcoin_tweets_clean.csv'


df_clean = pd.read_csv(tweets_clean_file,low_memory=False)
print(df_clean.shape)
df_clean.head(5)

# Drop na values
df_clean = df_clean.dropna()
# Manage data types
df_clean = df_clean.astype({"Likes": int, "Retweets": float, "Followers": int})
df_clean = df_clean.astype({"Retweets": int})

# Sentiment analysis

VADER (Valence Aware Dictionary and sEntiment Reasoner) is a lexicon and rule-based sentiment analysis tool that is specifically attuned to sentiments expressed in social media.

VADER takes into account 
- negations and contractions (not good, wasn’t good)
- Punctuation (good!!!), CAPS, emotes :), emojis 
- Intensificators (very, kind of), acronyms ‘lol’
- Scores between -1.0 (negative) and 1.0 (positive)

We will use this sentiment analysis of the tweets to calculate a score that will represent the importance of each tweet.

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

analyzer = SentimentIntensityAnalyzer()
compound = []
for i,s in enumerate(tqdm(df_clean['Tweets'])):
    vs = analyzer.polarity_scores(s)
    compound.append(vs["compound"])
df_clean["compound"] = compound
df_clean.head(5)

## Calculate a score for each tweet

To calculate the score for each tweet, we use different variables to which we had a weight based on its importance.

The compound column represents the sentiment of the tweets and its value is between -1 and 1.

We also use the number of retweets, the number of likes, and the number of users that follow the tweet's author.

import math

scores = []

for i, s in tqdm(df_clean.iterrows(), total=df_clean.shape[0]):
    scores.append(s["compound"] * ((int(s["Followers"]))) * ((int(s["Likes"])+1)) * ((int(s["Retweets"])+1)))

df_clean["score"] = scores
df_clean.head(20)

## Group by

# Group by hourly sum of scores
sent = df_clean.groupby('Hour')['score'].sum()
comp = df_clean.groupby('Hour')['compound'].sum()
n = df_clean.groupby('Hour')['Tweets'].count()
likes = df_clean.groupby('Hour')['Likes'].sum()
retweets = df_clean.groupby('Hour')['Retweets'].sum()

sent = pd.Series.to_frame(sent)
comp = pd.Series.to_frame(comp)
n = pd.Series.to_frame(n)
likes = pd.Series.to_frame(likes)
retweets = pd.Series.to_frame(retweets)

conc = pd.concat([n, likes, retweets, comp, sent], axis=1)

# Set index as datetime
conc.index = pd.to_datetime(conc.index)
# Sort by time
conc = conc.sort_index()

## Export to csv

twitter_sentiment_file = './data/twitter/bitcoin_twitter_sentiment.csv'

conc.to_csv(twitter_sentiment_file, mode='a', encoding='utf-8',index=True,header=True)
