==> 04_BitcoinPriceDataCollection.ipynb <==
# Bitcoin price collection

import numpy as np
import pandas as pd
import yfinance as yf 
import plotly.graph_objs as plot
import datetime
import time
import os

PRICE_FOLDER    = "data/price/"
TWITTER_FOLDER  = "data/twitter/"
tweets_raw_file = './data/twitter/bitcoin_tweets_raw.csv'
tweets_clean_file = './data/twitter/bitcoin_tweets_clean.csv'
price_raw_file = './data/price/bitcoin_price_raw.csv'
price_clean_file = './data/price/bitcoin_price_clean.csv'


Retrieve Bitcoin hourly price from Yahoo Finance

- Base price data is the first retrieval to a dataframe
- Updated data retrieves new hourly data to the same dataframe
- Data is exported to .csv after each updated retrieval

df_price_raw = pd.read_csv(price_raw_file,low_memory=False)

df_price_clean = pd.read_csv(price_clean_file,low_memory=False)

## Base price data

# Retrieve hourly price data
price_data = yf.download(tickers='BTC-USD', period='24h', interval='1h')
# Rename df index
price_data.index.name = 'Time'
# Head data
price_data.head(5)

# Drop last row of df
price_data = price_data[:-1]

## To .csv

# price_data to .csv
# Import same .csv to update

## Updated data

# Retrieve hourly price data
price_data_update = yf.download(tickers='BTC-USD', period='24h', interval='1h')
# Rename df index
price_data_update.index.name = 'Time'
# Drop last row
price_data_update = price_data_update[:-1]
# Head data
price_data_update.head(5)

## Merge data

# Concetenate original price data with updated data
price_data_live = pd.concat([price_data, price_data_update])
# Remove duplicate indices
price_data_live = price_data_live[~price_data_live.index.duplicated(keep='first')]

# Convert to .csv
price_data_live.to_csv(price_raw_file, encoding='utf-8')

## Plot data

#declare figure
fig = plot.Figure()

#Candlestick
fig.add_trace(plot.Candlestick(x=price_data_live.index,
                open=price_data_live['Open'],
                high=price_data_live['High'],
                low=price_data_live['Low'],
                close=price_data_live['Close'], name = 'market price data'))

# Add titles
fig.update_layout(
    title='Bitcoin live share price evolution',
    yaxis_title='Bitcoin Price (kUS Dollars)')

# X-Axes
fig.update_xaxes(
    rangeslider_visible=True,
    rangeselector=dict(
        buttons=list([
            dict(count=15, label="15m", step="minute", stepmode="backward"),
            dict(count=45, label="45m", step="minute", stepmode="backward"),
            dict(count=1, label="HTD", step="hour", stepmode="todate"),
            dict(count=6, label="6h", step="hour", stepmode="backward"),
            dict(step="all")
        ])
    )
)

#Show
fig.show()

!jupyter nbconvert --to script --no-prompt 04_BitcoinPriceDataCollection.ipynb
